requirements.txt
flask
transformers
pillow
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
--index-url https://download.pytorch.org/whl/cpu
flask-cors
